Modernize gem.
