+++ .editorconfig
@@ -3,3 +3,7 @@ root = true
 [*]
 indent_style = tab
 indent_size = 2
+
+[*.{yml,yaml}]
+indent_style = space
+indent_size = 2
