+++ .github/workflows/autobahn-server-tests.yaml
@@ -13,7 +13,7 @@ jobs:
     strategy:
       matrix:
         ruby:
-          - "3.2"
+          - "3.3"
         
         experimental: [false]
     
@@ -21,7 +21,7 @@ jobs:
       BUNDLE_WITH: autobahn_tests
     
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v4
     
     - uses: ruby/setup-ruby@v1
       with:
