+++ .github/workflows/test-external.yaml
@@ -20,12 +20,12 @@ jobs:
           - macos
         
         ruby:
-          - "3.0"
           - "3.1"
           - "3.2"
+          - "3.3"
     
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v4
     - uses: ruby/setup-ruby@v1
       with:
         ruby-version: ${{matrix.ruby}}
