+++ .github/workflows/test.yaml
@@ -21,9 +21,9 @@ jobs:
           - macos
         
         ruby:
-          - "3.0"
           - "3.1"
           - "3.2"
+          - "3.3"
         
         experimental: [false]
         
@@ -39,7 +39,7 @@ jobs:
             experimental: true
     
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v4
     - uses: ruby/setup-ruby@v1
       with:
         ruby-version: ${{matrix.ruby}}
