+++ protocol-websocket.gemspec
@@ -15,9 +15,14 @@ Gem::Specification.new do |spec|
 	
 	spec.homepage = "https://github.com/socketry/protocol-websocket"
 	
+	spec.metadata = {
+		"documentation_uri" => "https://socketry.github.io/protocol-websocket/",
+		"source_code_uri" => "https://github.com/socketry/protocol-websocket.git",
+	}
+	
 	spec.files = Dir.glob(['{lib}/**/*', '*.md'], File::FNM_DOTMATCH, base: __dir__)
 	
-	spec.required_ruby_version = ">= 3.0"
+	spec.required_ruby_version = ">= 3.1"
 	
 	spec.add_dependency "protocol-http", "~> 0.2"
 end
