+++ test/protocol/websocket/text_frame.rb
@@ -3,7 +3,7 @@
 # Released under the MIT License.
 # Copyright, 2019-2023, by Samuel Williams.
 
-require 'a_websocket_frame'
+require 'protocol/websocket/a_websocket_frame'
 require 'protocol/websocket/text_frame'
 
 describe Protocol::WebSocket::TextFrame do
@@ -20,7 +20,7 @@ describe Protocol::WebSocket::TextFrame do
 	with "with mask" do
 		let(:frame) {subject.new(mask: "abcd").pack("Hello World")}
 		
-		it_behaves_like AWebSocketFrame
+		it_behaves_like Protocol::WebSocket::AWebSocketFrame
 		
 		it "encodes binary representation" do
 			buffer = StringIO.new
@@ -34,7 +34,7 @@ describe Protocol::WebSocket::TextFrame do
 	with "without mask" do
 		let(:frame) {subject.new.pack("Hello World")}
 		
-		it_behaves_like AWebSocketFrame
+		it_behaves_like Protocol::WebSocket::AWebSocketFrame
 		
 		it "encodes binary representation" do
 			buffer = StringIO.new
